improved data scraping algorithm as well as updating names for TrackDetails and DriverDetails
