+++ src/components/DriverDetails/index.jsx
@@ -5,29 +5,33 @@ import * as d3 from 'd3'
 // Utils
 import useGlobalStore from '../../utils/store'
 import { useD3 } from '../../utils/useD3'
+import driverIdMapper from '../../utils/driverIdMapper'
 
-
-const RaceScrapeViz = ({raceList, scrapedRaceList}) => {
+const DriverDetails = ({data, scrapedDriverList}) => {
     const selectedRace = useGlobalStore((state) => state.selectedRace);
-    if (scrapedRaceList){
+    const hoveredDriverId = useGlobalStore(state => state.hoveredDriverId);
+    const selectedDrivers = useGlobalStore(state => state.selectedDrivers);
+
+    console.log(data);
+    if (selectedDrivers){
         return(
             <div>
                 <Grid>
-                    {raceList.map(function(race, i){
+                    {selectedDrivers.map(function(driver, i){
                         return (
                             <Grid.Col span={3}>
                                 <Card shadow="sm" p="md" radius="md" withBorder>
                                     <Card.Section>
                                         <Image
-                                        src={scrapedRaceList[race.circuitId].img_url}
-                                        alt={race.name}
+                                        src={scrapedDriverList[driver].img_url}
+                                        alt={driverIdMapper[driver].name}
                                         />
                                     </Card.Section>
                                     <Group position="apart" mt="md" mb="xs">
-                                        <Text weight={500}>{race.name}</Text>
+                                        <Text weight={500}>{driverIdMapper[driver].name}</Text>
                                     </Group>
                                     <Text size="sm" color="dimmed">
-                                        {scrapedRaceList[race.circuitId].description}
+                                        {scrapedDriverList[driver].description}
                                     </Text>
                                 </Card>
                             </Grid.Col>
@@ -39,4 +43,4 @@ const RaceScrapeViz = ({raceList, scrapedRaceList}) => {
     }
 }
 
-export default RaceScrapeViz
+export default DriverDetails
