+++ src/components/TrackDetails/index.jsx
@@ -5,33 +5,29 @@ import * as d3 from 'd3'
 // Utils
 import useGlobalStore from '../../utils/store'
 import { useD3 } from '../../utils/useD3'
-import driverIdMapper from '../../utils/driverIdMapper'
 
-const RaceScrapeViz = ({data, scrapedDriverList}) => {
+
+const TrackDetails = ({raceList, scrapedRaceList}) => {
     const selectedRace = useGlobalStore((state) => state.selectedRace);
-    const hoveredDriverId = useGlobalStore(state => state.hoveredDriverId);
-    const selectedDrivers = useGlobalStore(state => state.selectedDrivers);
-
-    console.log(data);
-    if (selectedDrivers){
+    if (scrapedRaceList){
         return(
             <div>
                 <Grid>
-                    {selectedDrivers.map(function(driver, i){
+                    {raceList.map(function(race, i){
                         return (
                             <Grid.Col span={3}>
                                 <Card shadow="sm" p="md" radius="md" withBorder>
                                     <Card.Section>
                                         <Image
-                                        src={scrapedDriverList[driver].img_url}
-                                        alt={driverIdMapper[driver].name}
+                                        src={scrapedRaceList[race.circuitId].img_url}
+                                        alt={race.name}
                                         />
                                     </Card.Section>
                                     <Group position="apart" mt="md" mb="xs">
-                                        <Text weight={500}>{driverIdMapper[driver].name}</Text>
+                                        <Text weight={500}>{race.name}</Text>
                                     </Group>
                                     <Text size="sm" color="dimmed">
-                                        {scrapedDriverList[driver].description}
+                                        {scrapedRaceList[race.circuitId].description}
                                     </Text>
                                 </Card>
                             </Grid.Col>
@@ -43,4 +39,4 @@ const RaceScrapeViz = ({data, scrapedDriverList}) => {
     }
 }
 
-export default RaceScrapeViz
+export default TrackDetails
